Update bookmark-plus to v0.1.0

+++ PKGBUILD
@@ -9,8 +9,8 @@ url="https://devan.gg/bookmark"
 license=('MIT')
 depends=()
 makedepends=('go')
-source=("${_binname}-${pkgver}.tar.gz::https://github.com/imdevan/bookmark/archive/refs/tags/v${pkgver}.tar.gz")
-sha256sums=('REPLACE_WITH_ACTUAL_SHA256')
+source=("${_binname}-${pkgver}.tar.gz::https://github.com/imdevan/bookmark/archive/refs/tags/v0.1.0.tar.gz")
+sha256sums=('c12a5e74fae7ef4d0c18c9f4d0516d8b7bf065e94e1a9abee0fbfbe03ae047f7')
 
 build() {
   cd "${_binname}-${pkgver}"
@@ -22,5 +22,7 @@ build() {
 package() {
   cd "${_binname}-${pkgver}"
   install -Dm755 ${_binname} "${pkgdir}/usr/bin/${_binname}"
-  install -Dm644 LICENSE "${pkgdir}/usr/share/licenses/${pkgname}/LICENSE"
+  if [ -f LICENSE ]; then
+    install -Dm644 LICENSE "${pkgdir}/usr/share/licenses/${pkgname}/LICENSE"
+  fi
 }
